+++ src/kanban/model/Task.java
@@ -3,7 +3,7 @@ package kanban.model;
 import java.time.LocalDateTime;
 
 public final class Task extends AbstractTask {
-    public Task(String title, String description, int id, LocalDateTime startTime, int duration) {
+    public Task(String title, String description, int id, LocalDateTime startTime, long duration) {
         super(title, description, id, startTime, duration);
     }
